Move fetch from gru
