+++ src/data/fetchFromGru.js
